+++ src/components/dropdown-suggest/index.js
@@ -358,7 +358,7 @@ class DropdownSuggest extends React.Component {
   get listHTML() {
     let listClasses = `${this.rootClass}__list` +
         (this.state.open ? '' : ' hidden') +
-        this.baseListClasses;
+        this.commonListClasses;
 
     return (
       <ul
@@ -371,6 +371,15 @@ class DropdownSuggest extends React.Component {
   }
 
   /**
+   * Extends the input content to include the input icon.
+   *
+   * @method additionalInputContent
+   */
+  get additionalInputContent() {
+    return this.inputIconHTML("dropdown");
+  }
+
+  /**
    * Renders the component.
    *
    * @method render
@@ -380,9 +389,8 @@ class DropdownSuggest extends React.Component {
       <div className={ this.mainClasses } >
 
         { this.labelHTML }
-        <input { ...this.inputProps } />
+        { this.inputHTML }
         <input { ...this.hiddenInputProps } />
-        { this.inputIconHTML("dropdown") }
         { this.validationHTML }
 
         { this.listHTML }